backend/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
django-filter==2.2.0
markdown==3.2.1
django-extensions==2.2.8
PyYAML==5.3
pytest-django==3.8.0
